+++ ChangeLog.md
@@ -23,6 +23,7 @@ Modelerfour version: 4.12.301
 **New Features**
 
 - Client side validation is now disabled by default  #558
+- We now also generate an async multiapi client when running multiapiscript  # 480
 
 ### 2020-04-06 - 5.0.0-dev.20200406.1
 Modelerfour version: 4.12.294